+++ src/JK.Common.Tests/Extensions/StringExtensionTests/ExtensionTests.cs
@@ -56,25 +56,6 @@ public class ExtensionTests
     }
 
     [Theory]
-    [InlineData("", 2, "")]
-    [InlineData("Test", 2, "st")]
-    [InlineData("Test", 4, "Test")]
-    [InlineData("Test", 6, "Test")]
-    public void Right_Theories(string input, int length, string expected)
-    {
-        var actual = input.Right(length);
-        Assert.Equal(expected, actual);
-    }
-
-    [Fact]
-    public void Reverse_Test()
-    {
-        const string s = "Sample Text";
-        var actual = s.Reverse();
-        Assert.Equal("txeT elpmaS", actual);
-    }
-
-    [Theory]
     [InlineData("$1,000,000.00", "1000000.00")]
     [InlineData("1000.00", "1000.00")]
     [InlineData("$,", "0")]
@@ -90,6 +71,35 @@ public class ExtensionTests
     public void RemoveXml_Test(string input, string expected)
     {
         var actual = input.RemoveXml();
+        Assert.Equal(expected, actual);
+    }
+
+    [Theory]
+    [InlineData("abc", "a", "bc")]
+    [InlineData("abc abc", "a", "bc bc")]
+    [InlineData("abc", "d", "abc")]
+    public void ReplaceWithEmpty_Theories(string input, string replace, string expected)
+    {
+        var actual = input.ReplaceWithEmpty(replace);
+        Assert.Equal(expected, actual);
+    }
+
+    [Fact]
+    public void Reverse_Test()
+    {
+        const string s = "Sample Text";
+        var actual = s.Reverse();
+        Assert.Equal("txeT elpmaS", actual);
+    }
+
+    [Theory]
+    [InlineData("", 2, "")]
+    [InlineData("Test", 2, "st")]
+    [InlineData("Test", 4, "Test")]
+    [InlineData("Test", 6, "Test")]
+    public void Right_Theories(string input, int length, string expected)
+    {
+        var actual = input.Right(length);
         Assert.Equal(expected, actual);
     }
 }